cleanup:  - removed unecessory wrapper component  - move earn to tsx  - updates styles import  - updated tooltip font sizes  - fix routeList import for Earn page
