+++ src/layout/routes/routeList.tsx
@@ -7,7 +7,7 @@ const Bridging = React.lazy(() => import('containers/Bridging'))
 const BobaScope = React.lazy(() => import('containers/bobaScope/BobaScope'))
 const OldDao = React.lazy(() => import('containers/dao/OldDao'))
 const DevTools = React.lazy(() => import('containers/devtools/DevTools'))
-const EarnWrapper = React.lazy(() => import('containers/earn/EarnWrapper'))
+const Earn = React.lazy(() => import('containers/earn/Earn'))
 const Home = React.lazy(() => import('containers/home'))
 const SaveWrapper = React.lazy(() => import('containers/save/SaveWrapper'))
 const History = React.lazy(() => import('containers/history/History'))
@@ -58,7 +58,7 @@ export const ROUTE_LIST: IRoute[] = [
       },
       {
         path: ROUTES_PATH.EARN,
-        element: <EarnWrapper />,
+        element: <Earn />,
         key: 'Earn',
       },
       {
